# Optimize `from_df()`