Depends on #93 

- [x] Vectorized date conversion. -> 13x impact for 0 Indexes, negligible for multi-index.
- [x] Very slight improvement by changing apply to map for checking mixed types in object columns. -> Barely noticeable 
- [x]  Remove `get_group(group)`. Call `group[column_name]` directly. -> 2x impact with multiple indexes, doesn't apply to no index.

>__Warning__: Had to add np.int64 to PYTHON_DTYPE_MAPPING in order to pass the test after implementing the third point. Don't really understand the reason for this change yet.  Check [this commit](https://github.com/google/temporian/pull/96/commits/8857a8f9b9947b4a326f4b5e5184aa22a8f6b463).

Previous time:
```
================================================================
Name                              Wall time (s)    CPU time (s)
================================================================
from_dataframe:s:10000_numidx:0_numidxval:20_idx:int       0.00198       0.00542
from_dataframe:s:10000_numidx:0_numidxval:20_idx:str       0.00193       0.00193
from_dataframe:s:10000_numidx:1_numidxval:20_idx:int       0.00529       0.00528
from_dataframe:s:10000_numidx:1_numidxval:20_idx:str       0.00897       0.00897
from_dataframe:s:10000_numidx:3_numidxval:20_idx:int       0.69247       0.69235
from_dataframe:s:10000_numidx:3_numidxval:20_idx:str       1.14818       1.13795
from_dataframe:s:10000_numidx:5_numidxval:20_idx:int       1.28118       1.27048
from_dataframe:s:10000_numidx:5_numidxval:20_idx:str       2.31480       2.30958
```

New time:

```
================================================================
Name                              Wall time (s)    CPU time (s)
================================================================
from_dataframe:s:10000_numidx:0_numidxval:20_idx:int       0.00015       0.00015
from_dataframe:s:10000_numidx:0_numidxval:20_idx:str       0.00014       0.00014
from_dataframe:s:10000_numidx:1_numidxval:20_idx:int       0.00192       0.00192
from_dataframe:s:10000_numidx:1_numidxval:20_idx:str       0.00545       0.00545
from_dataframe:s:10000_numidx:3_numidxval:20_idx:int       0.36431       0.36395
from_dataframe:s:10000_numidx:3_numidxval:20_idx:str       0.55161       0.55124
from_dataframe:s:10000_numidx:5_numidxval:20_idx:int       0.63044       0.63038
from_dataframe:s:10000_numidx:5_numidxval:20_idx:str       1.14939       1.14907
```

Speed up:

```
================================================================
Name                              Wall time Speed up  CPU time Speed up
================================================================
from_dataframe:s:10000_numidx:0_numidxval:20_idx:int       13.2x           36.1x
from_dataframe:s:10000_numidx:0_numidxval:20_idx:str       13.8x           13.8x
from_dataframe:s:10000_numidx:1_numidxval:20_idx:int        2.8x            2.8x
from_dataframe:s:10000_numidx:1_numidxval:20_idx:str        1.6x            1.6x
from_dataframe:s:10000_numidx:3_numidxval:20_idx:int        1.9x            1.9x
from_dataframe:s:10000_numidx:3_numidxval:20_idx:str        2.1x            2.1x
from_dataframe:s:10000_numidx:5_numidxval:20_idx:int        2.0x            2.0x
from_dataframe:s:10000_numidx:5_numidxval:20_idx:str        2.0x            2.0x
```

We can see from the table that the most important aspect to optimize is the logic when there are multiple indexes. Grouping probably takes the most time.

>__Warning__: Couldnt run the profiler to profile from_dataframe.